Refatoração da pasta 'cliente' para 'client' e exclusão de componentes que não serão utilizados.
