+++ client/README.md
@@ -1,4 +1,4 @@
-# Cliente
+# Client
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 15.2.3.
