+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cliente",
+  "name": "client",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
