+++ client/src/app/app.module.ts
@@ -7,7 +7,6 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { LoginComponent } from './login/login.component';
 import { RegisterComponent } from './cadastro/cadastro.component';
-import { EsqueciComponent } from './esqueci/esqueci.component';
 import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
@@ -16,7 +15,6 @@ import { HttpClientModule } from '@angular/common/http';
 
     LoginComponent,
     RegisterComponent,
-    EsqueciComponent,
   ],
   imports: [
     BrowserModule,
