+++ client/src/app/login/login.component.css
@@ -1,5 +1,150 @@
 @import url('https://fonts.googleapis.com/css2?family=Aldrich&family=Iceland&family=Inter:wght@400;600&family=Itim&family=Montserrat:wght@300;400;700&display=swap');
 
+/* RESET.CSS*/
+html,
+body,
+div,
+span,
+applet,
+object,
+iframe,
+h1,
+h2,
+h3,
+h4,
+h5,
+h6,
+p,
+blockquote,
+pre,
+a,
+abbr,
+acronym,
+address,
+big,
+cite,
+code,
+del,
+dfn,
+em,
+img,
+ins,
+kbd,
+q,
+s,
+samp,
+small,
+strike,
+strong,
+sub,
+sup,
+tt,
+var,
+b,
+u,
+i,
+center,
+dl,
+dt,
+dd,
+ol,
+ul,
+li,
+fieldset,
+form,
+label,
+legend,
+table,
+caption,
+tbody,
+tfoot,
+thead,
+tr,
+th,
+td,
+article,
+aside,
+canvas,
+details,
+embed,
+figure,
+figcaption,
+footer,
+header,
+hgroup,
+menu,
+nav,
+output,
+ruby,
+section,
+summary,
+time,
+mark,
+audio,
+video {
+    margin: 0;
+    padding: 0;
+    border: 0;
+    font-size: 100%;
+    font: inherit;
+    vertical-align: baseline;
+}
+
+/* HTML5 display-role reset for older browsers */
+article,
+aside,
+details,
+figcaption,
+figure,
+footer,
+header,
+hgroup,
+menu,
+nav,
+section {
+    display: block;
+}
+
+body {
+    line-height: 1;
+}
+
+ol,
+ul {
+    list-style: none;
+}
+
+blockquote,
+q {
+    quotes: none;
+}
+
+blockquote:before,
+blockquote:after,
+q:before,
+q:after {
+    content: '';
+    content: none;
+}
+
+table {
+    border-collapse: collapse;
+    border-spacing: 0;
+}
+
+/*RESET.CSS*/
+
+body {
+    width: 100%;
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    backdrop-filter: blur(2px);
+
+}
+
 .imagem-logo {
     margin-top: 2rem;
     margin-bottom: 1rem;
@@ -89,13 +234,13 @@ input::placeholder {
     height: 3rem;
     position: absolute;
     margin-top: 3rem;
-    left: 6.3rem;
+    left: 43rem;
 
     color: rgba(255, 255, 255, 0.6)
 }
 
 
-.esqueceu-senha {
+/*.esqueceu-senha {
     font-family: 'Inter', sans-serif;
     font-weight: 400;
     font-size: 18px;
@@ -108,7 +253,7 @@ input::placeholder {
     font-size: 20px;
     margin-left: 60%;
     padding-bottom: 1%;
-}
+}*/
 
 .entrar-botao button {
     background: #1A7FC1;
