+++ client/src/assets/logo_preta.png
