+++ client/src/styles.css
@@ -1,0 +1,3 @@
+body {
+    background-color: rgb(13, 37, 56);
+}
