+++ client/tsconfig.json
@@ -13,7 +13,6 @@
     "sourceMap": true,
     "declaration": false,
     "downlevelIteration": true,
-    "esModuleInterop": true,
     "experimentalDecorators": true,
     "moduleResolution": "node",
     "importHelpers": true,
