--- cliente/src/app/esqueci/esqueci.component.css
@@ -1,169 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Aldrich&family=Iceland&family=Inter:wght@400;600&family=Itim&family=Montserrat:wght@300;400;700&display=swap');
-
-
-.tela-login {
-    width: 680px;
-    height: 640px;
-    display: flex;
-    flex-direction: column;
-    background: #86929B;
-    border-radius: 32px;
-    align-items: center;
-}
-
-.imagem-logo {
-    margin-top: 2rem;
-    margin-bottom: 1rem;
-}
-
-.palavras {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.titulo {
-    font-family: 'Iceland', sans-serif;
-    font-size: 52px;
-    font-weight: bold;
-    margin-bottom: 2rem;
-}
-
-.input-box {
-    width: 440px;
-    display: block;
-}
-
-.cabeçalho {
-    width: 70%;
-    display: flex;
-    justify-content: center;
-    font-family: 'Inter', sans-serif;
-    font-size: 24px;
-    color: #1A1C2C;
-    padding-bottom: 0.3rem;
-}
-
-.esqueceu-palavra {
-    vertical-align: center;
-    align-items: center;
-    justify-content: center;
-    display: flex;
-    text-align: center;
-}
-
-.letra {
-    padding-bottom: 2rem;
-    text-align: center;
-}
-
-.input-box input {
-    width: 85%;
-    height: 3.2rem;
-    border: solid;
-    border-radius: 16px;
-    border-color: #000000;
-    background: rgba(13, 37, 56, 1);
-    margin-bottom: 1rem;
-    margin-top: 5px;
-
-    font-family: 'Itim', sans-serif;
-    font-style: normal;
-    font-weight: 400;
-    font-size: 24px;
-    color: rgba(255, 255, 255, 1);
-    vertical-align: center;
-    padding-left: 15%;
-}
-
-.input-box input:focus,
-.input-box input:hover {
-    outline: none;
-    border: solid;
-    border-color: rgba(255, 255, 255, 1);
-}
-
-input::placeholder {
-    font-family: 'Itim', sans-serif;
-    font-style: normal;
-    font-weight: 400;
-    font-size: 20px;
-    color: rgba(255, 255, 255, 0.6);
-    vertical-align: center;
-}
-
-.input-box label {
-    font-family: 'Iceland', sans-serif;
-    font-weight: 400;
-    font-size: 28px;
-    -webkit-text-stroke-width: 2px;
-    -webkit-text-stroke-color: rgba(0, 0, 0, 0.25);
-}
-
-.icones {
-    width: 2.5rem;
-    height: 2.5rem;
-    position: absolute;
-    margin-top: 2.6rem;
-    left: 8.6rem;
-
-    color: rgba(255, 255, 255, 0.6)
-}
-
-.esqueci-botao button {
-    background: #1A7FC1;
-    border-radius: 16px;
-    width: 100%;
-    height: 4rem;
-    margin-top: 1rem;
-    border: none;
-}
-
-.esqueci-botao :disabled {
-    background-color: rgba(0, 0, 0, 0.25);
-    cursor: standart;
-}
-
-.esqueci-botao button:enabled {
-    background: #1A7FC1;
-    cursor: pointer;
-}
-
-.esqueci-botao button:enabled:hover {
-    background: rgba(26, 127, 193, 0.7)
-}
-
-.esqueci-botao button {
-    font-family: 'Inter', sans-serif;
-    font-weight: 600;
-    font-size: 32px;
-    color: #E1E1E1;
-    text-decoration: none;
-}
-
-.validation-error div {
-    display: flex;
-    align-items: center;
-
-    margin-top: -3%;
-    margin-bottom: 0.1rem;
-}
-
-.validation-error div img {
-    width: 2rem;
-    height: 2rem;
-    padding-right: 0.5rem;
-    padding-left: 0.2rem;
-}
-
-.validation-error div p {
-    font-family: 'Aldrich', sans-serif;
-    font-style: normal;
-    color: #FF3737;
-    font-weight: 400;
-    font-size: 20px;
-}
-
-.validation-error {
-    border-color: #FF3737;
-}
