--- cliente/src/app/esqueci/esqueci.component.html
@@ -1,34 +1,0 @@
-<body>
-  <div class="tela-login">
-    <div class="logo-sfotipy">
-      <a href="#">
-        <img src="../../assets/logo_preta.png" class="imagem-logo" alt="logo-sfotipy">
-      </a>
-    </div>
-    <div class="palavras">
-      <h1 class="titulo">Redefinição de senha</h1>
-      <p class="cabeçalho esqueceu-palavra">Esqueceu sua senha?</p>
-      <p class="cabeçalho letra">Digite seu email que enviaremos um link para redefinir sua senha</p>
-    </div>
-    <form [formGroup]="EsqueciForm" (ngSubmit)="submit()" #formDir="ngForm">
-      <div class="input-box">
-        <label for="email" class="nomes-input">Insira seu E-mail</label>
-        <label for="email">
-          <img src="../../assets/email_branco.svg" class="icones" id="email-i" alt="icone de email">
-        </label>
-        <input id="email" type="email" name="email" placeholder="exemplo@gmail.com" formControlName="email">
-        <div class="validation-error">
-          <div *ngIf="(email.touched || submitted) && email.errors?.['required']"><img
-              src="../../assets/erro_input.svg">
-            <p>O email é obrigatório!</p>
-          </div>
-          <div *ngIf="email.touched && email.errors?.['email']"><img src="../../assets/erro_input.svg">
-            <p>Email inválido!</p>
-          </div>
-        </div>
-      </div>
-      <div class="esqueci-botao"><button type="submit" [disabled]="EsqueciForm.invalid">Enviar</button>
-      </div>
-    </form>
-  </div>
-</body>
