+++ server/usuarios/user.json
@@ -1,22 +1,1 @@
-[
-    {
-        "id": 1,
-        "name": "Angel",
-        "picture": "",
-        "email": "angel@gmail.com",
-        "password": "1234",
-        "followers": 0,
-        "artist": false,
-        "playlists": []
-    },
-    {
-        "id": 2,
-        "name": "Angel",
-        "picture": "",
-        "email": "angelwillyanfire@gmail.com",
-        "password": "1234",
-        "followers": 0,
-        "artist": false,
-        "playlists": []
-    }
-]
+[{"id":1,"name":"Angel","picture":"","email":"angel@gmail.com","password":"1234","followers":0,"artist":false,"playlists":[]},{"id":2,"name":"Angel","picture":"","email":"angelwillyanfire@gmail.com","password":"1234","followers":0,"artist":false,"playlists":[]},{"id":1,"name":"Angel willyan rodrigues de lima","picture":"","email":"angelwillyan@gmail.com","password":"1234","followers":0,"artist":false,"playlists":[]},{"id":1,"name":"Angel","picture":"","email":"angel_@gmail.com","password":"1234","followers":0,"artist":false,"playlists":[]},{"id":1,"name":"Angel","picture":"","email":"angelasfasf@gmail.com","password":"1234","followers":0,"artist":false,"playlists":[]},{"id":1,"name":"Angel","picture":"","email":"angel____@gmail.com","password":"1234","followers":0,"artist":false,"playlists":[]},{"id":1,"name":"Angel","picture":"","email":"angeasfasfl@gmail.com","password":"1234","followers":0,"artist":false,"playlists":[]},{"id":1,"name":"Angel","picture":"","email":"angel__@gmail.com","password":"1234","followers":0,"artist":false,"playlists":[]},{"id":1,"name":"Angel","picture":"","email":"afqasfqwf@gmail.com","password":"1234","followers":0,"artist":false,"playlists":[]},{"id":1,"name":"Angel","picture":"","email":"angel______@gmail.com","password":"1234","followers":0,"artist":false,"playlists":[]}]
